Moving app level migrations to its own file
